better (?) string concatenation
For reasons that I don’t fully understand the code wouldn’t compile on my Xcode 8.3 because of an “Ambiguous reference to member ‘+’” error in line 52.

I wonder if this is a bug in the current swift release.

So I rewrote the code a bit to join an array instead to create a multiline string. Hope you don’t mind.