Update README.md
Added reference to LiveCode is the direct descendent of MetaCard (and Runtime Revolution after that) which was originally a Linux clone of HyperCard. 
It's cross-platform and much more further developed than any other contender for heir to HC, it's script language is very much like HyperTalk! 
I often describe it as HyperCard 9.6!